style(worning): 消除了代码警告

+++ 3122004883/demo/src/main/java/utils/HammingUtils.java
@@ -28,7 +28,7 @@ public class HammingUtils {
         // 通过 simHash1 和 simHash2 获得它们的海明距离
         int distance = getHammingDistance(simHash1, simHash2);
         // 通过海明距离计算出相似度，并返回
-        return 0.01 * (100 - distance * 100 / 128);
+        return 0.01 * (100 - ((double) distance * 100) / 128);
     }
 
 }
